requirements.txt
absl-py==2.3.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.18
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
# Editable Git install with no remote (fidzulu-ml==0.1.0)
-e d:\run1_2026_fidzulu_module3\fidzulupythonml
flatbuffers==25.9.23
gast==0.7.0
google-pasta==0.2.0
greenlet==3.3.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.13.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
oracledb==3.4.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.45
stack-data==0.6.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tornado==6.5.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
wcwidth==0.2.14
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
